pyproject.toml
[project]
name = "lysqlnb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Leyton Addison-Roach", email = "git@laddisonroach.com" }
]
requires-python = ">=3.14"
dependencies = [
    "cyclopts>=4.5.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff==0.14.*",
]

[project.scripts]
lysqlnb = "lysqlnb:main"

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D1",
    "W191",
    "E111",
    "E114",
    "E117",
    "D206",
    "D300",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
    "TD",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "PLR2004",
    "S101",
    "S603",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 20
